Refactor: breaking apart large files with many definitions. Ficing linting isses.
