+++ src/Munin/Helpers/__init__.py
@@ -12,7 +12,7 @@ from .TreeSpecies import (
     PINUS_SYLVESTRIS,
     BETULA_PENDULA,
     BETULA_PUBESCENS,
-) #
+)
 
 # From Primitives.py
 from .Primitives import (
@@ -27,20 +27,19 @@ from .Primitives import (
     TopHeightDefinition,
     TopHeightMeasurement,
     QuadraticMeanDiameter,
-    Volume,
+    AtomicVolume,
+    CompositeVolume,
     AngleCount,
     AngleCountAggregator,
-    Tree, # Base class for trees
+    Tree,
     SingleTree,
     RepresentationTree,
-) #
+    SiteBase
+)
 
-# From Base.py
-from .Base import (
-    CircularPlot,
-    Stand,
-    # StandMetricAccessor is primarily for internal use within Stand properties
-) #
+from .plot import CircularPlot, StandMetricAccessor
+from .stand import Stand
+
 
 __all__ = [
     # TreeSpecies components
@@ -49,8 +48,9 @@ __all__ = [
     # Primitives components
     'Age', 'AgeMeasurement', 'Diameter_cm', 'Position', 'SiteIndexValue',
     'StandBasalArea', 'StandVolume', 'Stems', 'TopHeightDefinition',
-    'TopHeightMeasurement', 'QuadraticMeanDiameter', 'Volume',
+    'TopHeightMeasurement', 'QuadraticMeanDiameter', 'AtomicVolume',
+    'CompositeVolume',
     'AngleCount', 'AngleCountAggregator', 'Tree', 'SingleTree', 'RepresentationTree',
     # Base components
-    'CircularPlot', 'Stand',
+    'CircularPlot', 'Stand', 'StandMetricAccessor','SiteBase'
 ]
